[Edit] Updated Spearman Cards and Systems related to their Function
Updated movement and movemement Cost Calculation for Slows and Rooted
Updated Conditions for EntityModifier so they can support conditions on the affected target and conditions of the card used.
Updated Trigger Data and Trigger themself, to support reactions to the triggers before resolving the rest of the Effect
Updated MovementUtility with Movement that was on TilemapUtility
Updated how the Cost is calculated and Applied to support Slowing Effects.
Updated Targeting so Ai can use all TargetingModes
Updated EditorScripts to show the Data better
Updated Dragging for Cards and Characters as well as highlighting of tiles that can be targeted when dragging a card.